Improved courses API

+++ backend/app/Http/Controllers/CourseController.php
@@ -8,13 +8,19 @@ use Illuminate\Http\Request;
 class CourseController extends Controller
 {
     /**
+     * The request instance.
+     * @var Request
+     */
+    private $request;
+
+    /**
      * Create a new controller instance.
-     *
+     * @param Request $request
      * @return void
      */
-    public function __construct()
+    public function __construct(Request $request)
     {
-        //
+        $this->request = $request;
     }
 
     public function index()
@@ -25,6 +31,11 @@ class CourseController extends Controller
 
     public function create(Request $request)
     {
+        $this->validate($this->request, [
+            'name' => 'required',
+            'created_by' => 'required|numeric'
+        ]);
+
         try {
             $course = new Course;
             $course->name = $request->name;
@@ -34,12 +45,15 @@ class CourseController extends Controller
             $course->parent_id = $request->parent_id;
             $course->created_by = $request->created_by;
             $course->save();
-            return response()->json($course);
+
+            return response()->json([
+                'success' => 'Course created successfully',
+                'course' => $course
+            ], 201);
+
         } catch (\Throwable $e) {
             return response()->json([
-                'error' => [
-                    'description' => $e->getMessage()
-                ]
+                'error' => $e->getMessage()
             ], 500);
         }
     }
@@ -47,6 +61,11 @@ class CourseController extends Controller
     public function show($id)
     {
         $course = Course::find($id);
+        if(!$course) {
+            return response()->json([
+                'error' => 'Course does not exist.'
+            ], 404);
+        }
         return response()->json($course);
     }
 
@@ -54,21 +73,53 @@ class CourseController extends Controller
     {
         $course = Course::find($id);
 
-        $course->name = $request->input('name');
-        $course->description = $request->input('description');
-        $course->icon = $request->input('icon');
-        $course->slug = $request->input('slug');
-        $course->parent_id = $request->input('parent_id');
-        $course->created_by = $request->input('created_by');
+        if(!$course) {
+            return response()->json([
+                'error' => 'Course does not exist.'
+            ], 404);
+        }
 
-        $course->save();
-        return response()->json($course);
+        try {
+            $course->name = $request->input('name') ?: $course->name;
+            $course->description = $request->input('description') ?: $course->description;
+            $course->icon = $request->input('icon') ?: $course->icon;
+            $course->slug = $request->input('slug') ?: $course->slug;
+            $course->parent_id = $request->input('parent_id') ?: $course->parent_id;
+            $course->save();
+
+            return response()->json([
+                'success' => 'Course updated successfully',
+                'course' => $course
+            ], 201);
+        } catch (\Throwable $e) {
+            return response()->json([
+                'error' => $e->getMessage()
+            ], 500);
+        }
+
     }
 
     public function destroy($id)
     {
         $course = Course::find($id);
-        $course->delete();
-        return response()->json('course removed successfully');
+
+        if(!$course) {
+            return response()->json([
+                'error' => 'Course does not exist.'
+            ], 404);
+        }
+
+        try {
+            $course->delete();
+
+            return response()->json([
+                'success' => 'Course removed successfully',
+                'course' => $course
+            ], 202);
+        } catch (\Throwable $e) {
+            return response()->json([
+                'error' => $e->getMessage()
+            ], 500);
+        }
     }
 }
